feat: save and list provider.

+++ internal/service/bizconf/biz_config_service.go
@@ -2,15 +2,14 @@ package bizconf
 
 import (
 	"context"
-	"errors"
+	"fmt"
 
 	"github.com/JrMarcco/kuryr/internal/domain"
+	"github.com/JrMarcco/kuryr/internal/errs"
 	"github.com/JrMarcco/kuryr/internal/repository"
 )
 
 //go:generate mockgen -source=./biz_config_service.go -destination=./mock/biz_config_service.mock.go -package=bizconfmock -typed BizConfigService
-
-var ErrInvalidBizId = errors.New("[kuryr] biz id (owner_id) not set")
 
 type Service interface {
 	Save(ctx context.Context, bizConfig domain.BizConfig) error
@@ -26,16 +25,22 @@ type DefaultService struct {
 
 func (s *DefaultService) Save(ctx context.Context, bizConfig domain.BizConfig) error {
 	if bizConfig.Id == 0 {
-		return ErrInvalidBizId
+		return fmt.Errorf("%w: invalidate biz config id [ %d ]", errs.ErrInvalidParam, bizConfig.Id)
 	}
 	return s.repo.Save(ctx, bizConfig)
 }
 
 func (s *DefaultService) Delete(ctx context.Context, id uint64) error {
+	if id == 0 {
+		return fmt.Errorf("%w: invalidate biz config id [ %d ]", errs.ErrInvalidParam, id)
+	}
 	return s.repo.Delete(ctx, id)
 }
 
 func (s *DefaultService) GetById(ctx context.Context, id uint64) (domain.BizConfig, error) {
+	if id == 0 {
+		return domain.BizConfig{}, fmt.Errorf("%w: invalidate biz config id [ %d ]", errs.ErrInvalidParam, id)
+	}
 	return s.repo.GetById(ctx, id)
 }
